fixed index by embed

+++ skoots/lib/skeleton.py
@@ -156,9 +156,9 @@ def index_skeleton_by_embed(skeleton: Tensor, embed: Tensor) -> Tensor:
     b, c, x, y, z = embed.shape                     # get the shape of the embedding
     embed = embed.view((c, -1)).round()             # flatten the embedding to extract it as an index
 
-    x_ind = embed[0, :].clamp(0, x)                 # Use the embedding as an x,y,z index
-    y_ind = embed[1, :].clamp(0, y)
-    z_ind = embed[2, :].clamp(0, z)
+    x_ind = embed[0, :].clamp(0, x).long()          # Use the embedding as an x,y,z index
+    y_ind = embed[1, :].clamp(0, y).long()
+    z_ind = embed[2, :].clamp(0, z).long()
 
     b, c, x, y, z = skeleton.shape
     out = torch.zeros_like(skeleton).flatten()      # For indexing to work, the out tensor has to be flat
@@ -167,11 +167,6 @@ def index_skeleton_by_embed(skeleton: Tensor, embed: Tensor) -> Tensor:
     out[ind] = skeleton[:, :, x_ind, y_ind, z_ind]  # assign the skeleton ind to the out tensor
 
     return out.view(b, c, x, y, z)                  # return the re-shaped out tensor
-
-
-
-
-
